fastapi_app/requirements.txt
anyio==4.9.0
asgiref==3.5.0
fastapi==0.115.13
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
email_validator==2.2.0
SQLAlchemy==2.0.41
aiomysql==0.2.0
asyncmy==0.2.10
aioredis==2.0.1
alembic==1.15.2
passlib==1.7.4
requests==2.32.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
